refactor(chat): split context.py 411 lines into focused modules
Per CLAUDE.md (single file <300 lines + single responsibility):
- chat/_context_images.py: download permission + image collection pipeline
- chat/_context_messages.py: LLM message list builders (group / p2p)
- chat/context.py: ChatContext dataclass + build_chat_context orchestrator + ContextVars

Public import path 'from app.chat.context import ...' unchanged.

Spec: docs/superpowers/specs/2026-05-07-dataflow-phase-6-cleanup-design.md §3.1
