+++ apps/agent-service/app/chat/context.py
@@ -1,34 +1,36 @@
 """Chat context builder — assemble LLM messages from DB history.
 
-Responsibilities:
+Responsibilities (orchestrator):
   - Fetch recent history via quick_search
-  - Process images (cached TOS URL / full pipeline)
-  - Build group or P2P message lists with image registry
   - Detect proactive scan messages and extract stimulus
+  - Delegate image collection to _context_images.collect_images
+  - Delegate message-shape construction to _context_messages.build_*
+  - Emit ConversationMessageContentSynced for any new TOS files
 
 Image pipeline:
   1. feishu image_key -> process_image -> TOS URL
   2. TOS URL -> ImageRegistry.register -> N.png
   3. text references images as @N.png
+
+Sub-modules (private to chat/):
+  - _context_images.py: download permission cache + image collection
+  - _context_messages.py: group / p2p LLM message list builders
 """
 
 from __future__ import annotations
 
-import asyncio
 import contextvars
 import logging
-import time
 from dataclasses import dataclass
 
 from langchain_core.messages import AIMessage, HumanMessage
 
-from app.agent.prompts import get_prompt
+from app.chat._context_images import collect_images
+from app.chat._context_messages import build_group_messages, build_p2p_messages
 from app.chat.content_parser import parse_content
-from app.chat.quick_search import QuickSearchResult, quick_search
-from app.data.queries import find_group_download_permission
-from app.data.session import get_session
+from app.chat.quick_search import quick_search
 from app.domain.chat_events import ConversationMessageContentSynced
-from app.infra.image import ImageRegistry, image_client
+from app.infra.image import ImageRegistry
 from app.infra.redis import get_redis
 from app.runtime import emit
 
@@ -43,10 +45,6 @@ is_proactive_var: contextvars.ContextVar[bool] = contextvars.ContextVar(
 proactive_stimulus_var: contextvars.ContextVar[str] = contextvars.ContextVar(
     "proactive_stimulus", default=""
 )
-
-# Download permission cache: chat_id -> (allows, expire_monotonic)
-_download_cache: dict[str, tuple[bool, float]] = {}
-_CACHE_TTL = 600  # 10 min
 
 
 @dataclass
@@ -112,7 +110,7 @@ async def build_chat_context(
     proactive_stimulus_var.set(proactive_stimulus)
 
     # --- Image processing ---
-    image_key_to_url, image_key_to_file = await _collect_images(l1_results, chat_type)
+    image_key_to_url, image_key_to_file = await collect_images(l1_results, chat_type)
 
     # Persist new TOS files in background via dataflow (Phase 6 v4 Gap 5).
     if image_key_to_file:
@@ -152,11 +150,11 @@ async def build_chat_context(
 
     # Build messages
     if chat_type == "group":
-        messages = _build_group_messages(
+        messages = build_group_messages(
             l1_results, effective_trigger_id, image_key_to_url, image_key_to_filename
         )
     else:
-        messages = _build_p2p_messages(
+        messages = build_p2p_messages(
             l1_results,
             image_key_to_url,
             image_key_to_filename,
@@ -179,233 +177,3 @@ async def build_chat_context(
         chat_name=chat_name,
         chain_user_ids=chain_user_ids,
     )
-
-
-# ---------------------------------------------------------------------------
-# Image collection
-# ---------------------------------------------------------------------------
-
-
-async def _allows_download(chat_id: str, chat_type: str) -> bool:
-    """Check group download permission with in-memory cache."""
-    if chat_type == "p2p":
-        return True
-
-    now = time.monotonic()
-    cached = _download_cache.get(chat_id)
-    if cached and cached[1] > now:
-        return cached[0]
-
-    try:
-        async with get_session() as s:
-            setting = await find_group_download_permission(s, chat_id)
-        allows = setting != "not_anyone"
-    except Exception:
-        logger.warning(
-            "Download permission check failed for %s, defaulting to allow", chat_id
-        )
-        allows = True
-
-    _download_cache[chat_id] = (allows, now + _CACHE_TTL)
-    return allows
-
-
-async def _collect_images(
-    results: list[QuickSearchResult],
-    chat_type: str,
-) -> tuple[dict[str, str], dict[str, str]]:
-    """Collect image URLs from message history.
-
-    Returns (image_key -> url, image_key -> tos_file_name).
-    """
-    cached_keys: list[tuple[str, str]] = []  # (image_key, tos_file)
-    uncached_keys: list[tuple[str, str, str]] = []  # (image_key, message_id, role)
-
-    for msg in results:
-        parsed = parse_content(msg.content)
-        for key in parsed.image_keys:
-            if key.startswith("@"):
-                continue
-            tos_file = parsed.tos_files.get(key)
-            if tos_file:
-                cached_keys.append((key, tos_file))
-            else:
-                uncached_keys.append((key, msg.message_id, msg.role))
-
-    image_key_to_url: dict[str, str] = {}
-    image_key_to_file: dict[str, str] = {}
-
-    # Cached images: sign URLs only
-    if cached_keys:
-        tasks = [image_client.get_url(tos_file) for _, tos_file in cached_keys]
-        url_results = await asyncio.gather(*tasks, return_exceptions=True)
-        for i, result in enumerate(url_results):
-            key, tos_file = cached_keys[i]
-            if isinstance(result, str) and result:
-                image_key_to_url[key] = result
-                image_key_to_file[key] = tos_file
-            else:
-                uncached_keys.append((key, "", ""))
-                logger.warning("TOS URL sign failed, falling back: %s", key)
-
-    # Permission check before downloading
-    if uncached_keys:
-        chat_id = results[0].chat_id or ""
-        if not await _allows_download(chat_id, chat_type):
-            logger.info(
-                "Group %s disallows download, skipping %d images",
-                chat_id,
-                len(uncached_keys),
-            )
-            uncached_keys = []
-
-    # Uncached: full pipeline (Lark download -> compress -> TOS)
-    if uncached_keys:
-        tasks = [
-            image_client.process_image(key, msg_id if role == "user" else None)
-            for key, msg_id, role in uncached_keys
-        ]
-        process_results = await asyncio.gather(*tasks, return_exceptions=True)
-        for i, result in enumerate(process_results):
-            key, msg_id, _ = uncached_keys[i]
-            if isinstance(result, dict) and result:
-                image_key_to_url[key] = result["url"]
-                if result.get("file_name"):
-                    image_key_to_file[key] = result["file_name"]
-            else:
-                logger.warning("Image processing failed: key=%s, msg=%s", key, msg_id)
-
-    return image_key_to_url, image_key_to_file
-
-
-# ---------------------------------------------------------------------------
-# Message list builders
-# ---------------------------------------------------------------------------
-
-
-def _image_fn(
-    image_key_to_filename: dict[str, str],
-):
-    """Return an image render function for parse_content.render()."""
-
-    def _fn(_i: int, key: str) -> str:
-        fn = image_key_to_filename.get(key)
-        return f"@{fn}" if fn else "[图片]"
-
-    return _fn
-
-
-def _extract_reply_chain(
-    messages: list[QuickSearchResult], trigger_id: str
-) -> tuple[list[QuickSearchResult], list[QuickSearchResult]]:
-    """Trace reply_message_id chain from trigger upward.
-
-    Returns (chain_messages, other_messages), both in time-ascending order.
-    """
-    msg_map = {msg.message_id: msg for msg in messages}
-    chain_ids: set[str] = set()
-    current_id: str | None = trigger_id
-
-    while current_id and current_id in msg_map:
-        chain_ids.add(current_id)
-        current_id = msg_map[current_id].reply_message_id
-
-    chain = [m for m in messages if m.message_id in chain_ids]
-    other = [m for m in messages if m.message_id not in chain_ids]
-    return chain, other
-
-
-def _build_group_messages(
-    messages: list[QuickSearchResult],
-    trigger_id: str,
-    image_key_to_url: dict[str, str],
-    image_key_to_filename: dict[str, str],
-) -> list[HumanMessage | AIMessage]:
-    """Build group chat message list.
-
-    Reply chain messages include image content blocks; other messages
-    reference images as @N.png in text only.
-    """
-    chain, other = _extract_reply_chain(messages, trigger_id)
-    img_fn = _image_fn(image_key_to_filename)
-
-    chain_lines = []
-    for msg in chain:
-        time_str = msg.create_time.strftime("%H:%M:%S")
-        username = msg.username or "未知用户"
-        text = parse_content(msg.content).render(image_fn=img_fn)
-        marker = " ⭐" if msg.message_id == trigger_id else ""
-        chain_lines.append(f"[{time_str}] {username}: {text}{marker}")
-
-    other_lines = []
-    for msg in other:
-        time_str = msg.create_time.strftime("%H:%M:%S")
-        username = msg.username or "未知用户"
-        text = parse_content(msg.content).render(image_fn=img_fn)
-        other_lines.append(f"[{time_str}] {username}: {text}")
-
-    user_content = get_prompt("context_builder").compile(
-        reply_chain="\n".join(chain_lines) if chain_lines else "（无回复链）",
-        other_messages="\n".join(other_lines) if other_lines else "（无其他消息）",
-    )
-
-    content_blocks: list = [{"type": "text", "text": user_content}]
-
-    # Attach reply chain images as content blocks
-    for msg in chain:
-        parsed = parse_content(msg.content)
-        for key in parsed.image_keys:
-            fn = image_key_to_filename.get(key)
-            url = image_key_to_url.get(key)
-            if fn and url:
-                content_blocks.append({"type": "text", "text": f"@{fn}:"})
-                content_blocks.append({"type": "image", "url": url})
-
-    return [HumanMessage(content_blocks=content_blocks)]  # type: ignore[arg-type]
-
-
-def _build_p2p_messages(
-    messages: list[QuickSearchResult],
-    image_key_to_url: dict[str, str],
-    image_key_to_filename: dict[str, str],
-    current_persona_id: str = "",
-) -> list[HumanMessage | AIMessage]:
-    """Build P2P message list with full image content blocks."""
-    result: list[HumanMessage | AIMessage] = []
-    img_fn = _image_fn(image_key_to_filename)
-
-    for msg in messages:
-        parsed = parse_content(msg.content)
-        text_content = parsed.render(image_fn=img_fn)
-
-        content_blocks: list = []
-        if text_content:
-            content_blocks.append({"type": "text", "text": text_content})
-
-        for key in parsed.image_keys:
-            fn = image_key_to_filename.get(key)
-            url = image_key_to_url.get(key)
-            if fn and url:
-                content_blocks.append({"type": "text", "text": f"@{fn}:"})
-                content_blocks.append({"type": "image", "url": url})
-            elif not fn:
-                logger.warning(
-                    "Image not registered: key=%s, msg=%s", key, msg.message_id
-                )
-
-        if not content_blocks:
-            continue
-
-        # Current persona's messages -> AIMessage; everything else -> HumanMessage
-        msg_persona_id = getattr(msg, "persona_id", None)
-        is_self = (
-            msg.role == "assistant"
-            and bool(msg_persona_id)
-            and msg_persona_id == current_persona_id
-        )
-        if is_self:
-            result.append(AIMessage(content_blocks=content_blocks))  # type: ignore[arg-type]
-        else:
-            result.append(HumanMessage(content_blocks=content_blocks))  # type: ignore[arg-type]
-
-    return result
